Use joint-js text wrapping instead of our own
In order to prepare for new <cmmn:documentation> element.
The previous text wrapping implementation did a sort of wild guess on the name attribute, if it contains newlines it would start create a new name without newlines and store the name with newlines in the description field, which is rather confusing and complex behavior.
Additionally we had confusing construts on the combination of name and id. This is also simply removed.

The old method of formatting to do custom text wrapping with "your own" newlines through description field is still possible, but is handled much more explicitly.