Decouple joint HTML id from definition id
If the definition for a joint element changes (this can happen in a CaseFileItem shape), the definition id can no longer be used to find the html (as that was created with the original id).
Fix that by making it a separately isolated variable that defaults to the initial definition's id.