Smarter hovering over elements
When hovering over the text inside an svg element, pointerout got triggered on the surrounding svg element, even though the pointer was still inside the element.
This led to confusing behavior when hovering e.g. an entry criterion on a human task and touching the text of the task (it then erroneously grayed out the entry criterion image)