Enable moving criteria on bounding box
The whole bounding box of the shape can now be selected to move the sentry shape around the edges of the parent shape.
E.g. When you try to move the flash or diamonds and you don't need to point mouse exactly withing the small area of the diamond or flash shape.