Increase numbers to avoid unexpected migrations
Instead of reusing "free" numbers for a certain type now always generated a new number, larger than any existing one.
This way we can avoid that someone wants to migrate a definition by adjusting an existing model with the following scenario:
- remove one task, add an entirely different new task.

With reusing the free number, the case engine assumes the existing task is being migrated, instead of dropped and a new one created.
To understand the background of such a scenario seems somewhat too much of a challenge ...