Refactor parsing and drop repository.ide connection
When the repository content is received, it is immediately parsed, instead of on demand.
It seems to be fast enough, and helps to implement refactor actions across models.

Also Repository class no longer needs the IDE pointer for warnings. This is done through changing the callback structures to use the Followup infrastructure.