Improve selecting mapping parameters
Instead of searching the entire case definition, just search in the task itself, as that needs to have the parameter.
Note: this now goes only by id, no longer by name (which should anyway be the default, as we're inside the task definition itself)

Also: dropped ImplementationParameterDefinition and replace it with plain ParameterDefinition, as it fits the same need.