Refactor dragdata dependencies
DragData is now standalone, and no longer depends on the IDE.
Whether dragging is done currently (needed inside CaseView) is now available as a static on DragData.
Also made more functionality specific logic in the subclasses of DragData and their usage.

This also reduces a "BIG TODO HERE" on the ElementRegistry initialization