Change server file to keep definition
When the source of the serverfile is set to a new version, then it would delete the existing definition.
This leads to unclear connections between serverfile references (especially when it is at the level of individual elements somewhere inside the file definition).
This behavior is now removed, and can only be invoked by explicitly calling file.clear().

Since file.clear() also clears the used references, the case model editor need not also clear the dimensions file when refreshing.