ImportNode creation always done in ModelDefinition
ModelDefinition must have an XMLDocument, therefore this always has an import node.
When an empty CaseFileItemView is created, it is not possible to draw a line from it, as the underlying definition does not have an importNode (yet).
This is due to the creation of ElementDefinitions inside ModelDefinition.createDefinition with an undefined importNode (which is bad code, but not fixed here).
Basically the code change in this commit works around that issue.