Refactor TypeRenderer class
Introduced generic TypeDefinitionElement in order to serve as a base class for SchemaDefinition and SchemaPropertyDefinition.

Made TypeRenderer a generic on TypeDefinitionElement, and separated PropertyRenderer and SchemaRenderer in their own file.

Now we no longer have an error when loading the bundle with the separate files

Also changed field declarations here and there

also updated collapse logic to have more understandable values