Fix editing name when creating series of tasks
When creating series of new elements from the ShapeBox, for each new element properties appear, and we can type a more functional name for the element.
Then, after typing we immediately drag a new element, it did not capture that name change, because the new element was not properly selected.
This is fixed.